[Feat/#14] 파일 분리
