+++ HMH_iOS/HMH_iOS/Global/Extension/UIImage+.swift
@@ -1,0 +1,24 @@
+//
+//  UIImage+.swift
+//  HMH_iOS
+//
+//  Created by 지희의 MAC on 1/5/24.
+//
+
+import UIKit
+
+extension UIImage {
+    static func load(named imageName: String) -> UIImage {
+        guard let image = UIImage(named: imageName, in: nil, compatibleWith: nil) else {
+            return UIImage()
+        }
+        return image
+    }
+    
+    func resize(to size: CGSize) -> UIImage {
+        let image = UIGraphicsImageRenderer(size: size).image { _ in
+            draw(in: CGRect(origin: .zero, size: size))
+        }
+        return image
+    }
+}
