+++ HMH_iOS/HMH_iOS/Global/Literals/Literal/Image.swift
@@ -22,18 +22,3 @@ enum ImageLiterals {
     }
 }
 
-extension UIImage {
-    static func load(named imageName: String) -> UIImage {
-        guard let image = UIImage(named: imageName, in: nil, compatibleWith: nil) else {
-            return UIImage()
-        }
-        return image
-    }
-    
-    func resize(to size: CGSize) -> UIImage {
-        let image = UIGraphicsImageRenderer(size: size).image { _ in
-            draw(in: CGRect(origin: .zero, size: size))
-        }
-        return image
-    }
-}
