--- HMH_iOS/HMH_iOS/Global/Literals/String/E.swift
@@ -1,8 +1,0 @@
-//
-//  E.swift
-//  HMH_iOS
-//
-//  Created by Seonwoo Kim on 12/23/23.
-//
-
-import Foundation
