+++ HMH_iOS/HMH_iOS/Global/Literals/String/String.swift
@@ -1,0 +1,16 @@
+//
+//  String.swift
+//  HMH_iOS
+//
+//  Created by 지희의 MAC on 1/5/24.
+//
+
+import Foundation
+
+enum StringLiteral {
+    enum TabBar {
+        static var challengeTitle = "챌린지"
+        static var homeTitle = "홈"
+        static var myPageTitle = "마이페이지"
+    }
+}
